+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 - Added method `Bitrix24AccountRepositoryInterface::findByApplicationToken` in contracts for
   support «[Delete Application](https://github.com/bitrix24/b24phpsdk/issues/62)» use case
 
+### Changed
+
+- Added nullable argument `bitrix24UserId` in method `Bitrix24AccountRepositoryInterface::findByMemberId` in contracts
+  for support use case «[RenewAuthToken](https://github.com/bitrix24/b24phpsdk/issues/63)»
+
 ### Fixed
 
 - Fixed errors in `Bitrix24\SDK\Services\Workflows\Common\WorkflowDocumentId`,